// services/webapp/src/sdk/models/CompetitionsPage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Open Dossard
 * Documentation de l\'API Open Dossard
 *
 * The version of the OpenAPI document: 1.0
 * Contact: contact@opendossard.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    CompetitionEntity,
    CompetitionEntityFromJSON,
    CompetitionEntityFromJSONTyped,
    CompetitionEntityToJSON,
} from './';

/**
 * 
 * @export
 * @interface CompetitionsPage
 */
export interface CompetitionsPage {
    /**
     * 
     * @type {Array<CompetitionEntity>}
     * @memberof CompetitionsPage
     */
    data?: Array<CompetitionEntity>;
    /**
     * 
     * @type {number}
     * @memberof CompetitionsPage
     */
    page?: number;
    /**
     * 
     * @type {number}
     * @memberof CompetitionsPage
     */
    totalCount?: number;
}

export function CompetitionsPageFromJSON(json: any): CompetitionsPage {
    return CompetitionsPageFromJSONTyped(json, false);
}

export function CompetitionsPageFromJSONTyped(json: any, ignoreDiscriminator: boolean): CompetitionsPage {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'data': !exists(json, 'data') ? undefined : ((json['data'] as Array<any>).map(CompetitionEntityFromJSON)),
        'page': !exists(json, 'page') ? undefined : json['page'],
        'totalCount': !exists(json, 'totalCount') ? undefined : json['totalCount'],
    };
}

export function CompetitionsPageToJSON(value?: CompetitionsPage | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'data': value.data === undefined ? undefined : ((value.data as Array<any>).map(CompetitionEntityToJSON)),
        'page': value.page,
        'totalCount': value.totalCount,
    };
}